Extract functions to a lib file
