+++ config.ts
@@ -2,7 +2,7 @@ import axios, { AxiosInstance } from "axios";
 
 const axiosClient = (baseUrl: string): AxiosInstance => {
   if (!baseUrl) {
-    throw new Error("Base url required to instantiate the axios client");
+    throw new Error("Base url required to instantiate axios client");
   }
   return axios.create({
     baseURL: baseUrl,
